{"schema":"decodex/commit/1","summary":"Move live click capture target to owner","authority":"manual"}
