+++ packages/rsnap-overlay/src/overlay/input_runtime.rs
@@ -8,6 +8,7 @@ use winit::keyboard::ModifiersState;
 
 #[cfg(target_os = "macos")]
 use crate::overlay::FrozenGlobalHotkey;
+use crate::overlay::live_capture_target::LiveClickCaptureTarget;
 use crate::overlay::runtime_timing::CURSOR_EVENT_TICK_TTL;
 #[cfg(target_os = "macos")]
 use crate::overlay::runtime_timing::SLOW_OP_WARN_CURSOR_LOCATION;
@@ -16,9 +17,9 @@ use crate::overlay::toolbar_layout_model;
 use crate::overlay::{
 	CursorMoveTrace, DeviceCursorPointSource, ElementState, FrozenSelectionDragCursorMoveTiming,
 	FrozenTextEditState, FrozenTextInputSource, FrozenToolbarTool, GlobalPoint, Ime, Key,
-	LiveCaptureInteraction, LiveClickCaptureTarget, Modifiers, MonitorRect, MouseScrollDelta,
-	NamedKey, OverlayControl, OverlayKeyboardInputEvent, OverlayMode, OverlaySession,
-	PhysicalPosition, PhysicalSize, PngAction, Pos2, Vec2, WindowId, WindowRenderer,
+	LiveCaptureInteraction, Modifiers, MonitorRect, MouseScrollDelta, NamedKey, OverlayControl,
+	OverlayKeyboardInputEvent, OverlayMode, OverlaySession, PhysicalPosition, PhysicalSize,
+	PngAction, Pos2, Vec2, WindowId, WindowRenderer,
 };
 
 impl OverlaySession {
