+++ packages/rsnap-overlay/src/overlay/tests/self_capture_runtime.rs
@@ -5,11 +5,11 @@ use std::ptr;
 use image::{Rgba, RgbaImage};
 
 #[cfg(target_os = "macos")]
-use crate::overlay::LiveClickCaptureTarget;
-#[cfg(target_os = "macos")]
 use crate::overlay::OverlayMode;
 #[cfg(target_os = "macos")]
 use crate::overlay::RectPoints;
+#[cfg(target_os = "macos")]
+use crate::overlay::live_capture_target::LiveClickCaptureTarget;
 #[cfg(target_os = "macos")]
 use crate::overlay::runtime_timing::DISPLAY_FIRST_FREEZE_LIVE_TIMEOUT;
 #[cfg(target_os = "macos")]
